requirements.txt
gspread
oauth2client
flask
flask-cors
requests
uuid
datetime
twilio
google-cloud-dialogflow
python-dotenv